requirements.txt
alabaster==0.7.12
aniso8601==8.0.0
antlr4-python3-runtime==4.9
appnope==0.1.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
biolinkml==1.4.9
biolink-model==1.2.5
bleach==3.1.1
bmt==0.1.1
cachetools==4.0.0
certifi==2019.11.28
CFGraph==0.2.1
chardet==3.0.4
click==7.1.1
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
entrypoints==0.3
flasgger==0.9.4
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
graphviz==0.13.2
gunicorn==20.0.4
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
incremental==17.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jsonasobj==1.2.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.2.0
mypy==0.770
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.4
neo4j==1.7.6
neo4jrestclient==2.1.1
neobolt==1.7.16
neotime==1.7.4
networkx==2.4
notebook==6.0.3
numpy==1.18.1
ordered-set==4.0.2
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
pbr==5.4.4
pexpect==4.8.0
pickleshare==0.7.5
pip-check-reqs==2.0.3
pluggy==0.13.1
prefixcommons==0.1.9
prologterms==0.0.6
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
py==1.8.1
Pygments==2.6.1
PyJSG==0.11.6
pyparsing==2.4.6
pyrsistent==0.15.7
PyShEx==0.7.14
PyShExC==0.8.2
pystache==0.5.4
pytest==5.4.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
rdflib==5.0.0
rdflib-jsonld==0.5.0
redis==3.4.1
requests==2.22.0
requests-cache==0.5.2
Send2Trash==1.5.0
ShExJSG==0.7.1
six==1.14.0
snowballstemmer==2.0.0
sparql-slurper==0.3.3
SPARQLWrapper==1.8.5
Sphinx==2.4.4
sphinx-click==2.3.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
stringcase==1.2.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.7
validators==0.14.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
zipp==3.1.0
jsonlines>=1.2.0
